+++ pgvector/peewee/vector.py
@@ -16,8 +16,9 @@ class VectorField(Field):
     def db_value(self, value: object) -> str | None:
         return Vector._to_db(value)
 
-    def python_value(self, value: Any) -> Vector | None:
-        return Vector._from_db(value)
+    def python_value(self, value: Any) -> list[float] | None:
+        value = Vector._from_db(value)
+        return None if value is None else value.to_list()
 
     def _distance(self, op: str, vector: object) -> Expression:
         return Expression(lhs=self, op=op, rhs=self.to_value(vector))