+++ pgvector/sqlalchemy/vector.py
@@ -31,8 +31,9 @@ class VECTOR(UserDefinedType[Any]):
         return process
 
     def result_processor(self, dialect: Dialect, coltype: Any) -> Any:
-        def process(value: Any) -> Vector | None:
-            return Vector._from_db(value)
+        def process(value: Any) -> list[float] | None:
+            value = Vector._from_db(value)
+            return None if value is None else value.to_list()
         return process
 
     class Comparator(TypeEngine.Comparator[Any]):